Refactor llamacpp plugin
