+++ src-tauri/plugins/tauri-plugin-llamacpp/src/utils/network.rs
@@ -1,0 +1,31 @@
+use std::collections::HashSet;
+use rand::{rngs::StdRng, Rng, SeedableRng};
+
+/// Check if a port is available for binding
+pub fn is_port_available(port: u16) -> bool {
+    std::net::TcpListener::bind(("127.0.0.1", port)).is_ok()
+}
+
+/// Generate a random port that's not in the used_ports set and is available
+pub fn generate_random_port(used_ports: &HashSet<u16>) -> Result<u16, String> {
+    const MAX_ATTEMPTS: u32 = 20000;
+    let mut attempts = 0;
+    let mut rng = StdRng::from_entropy();
+
+    while attempts < MAX_ATTEMPTS {
+        let port = rng.gen_range(3000..4000);
+
+        if used_ports.contains(&port) {
+            attempts += 1;
+            continue;
+        }
+
+        if is_port_available(port) {
+            return Ok(port);
+        }
+
+        attempts += 1;
+    }
+
+    Err("Failed to find an available port for the model to load".into())
+}
